feat: add noti when delete collection

+++ frontend/src/components/client/CollectionManagementPage/DeleteRecord.jsx
@@ -1,12 +1,39 @@
 import { Button } from 'antd';
 import { deleteCollection } from '../../../services/collections';
 import { DeleteOutlined } from '@ant-design/icons';
+import Swal from 'sweetalert2';
 
 const DeleteRecord = ({ collection, onReload }) => {
-  const handleDelete = async () => {
-    await deleteCollection(collection.id);
-    onReload();
-  }
+  const handleDelete = () => {
+    Swal.fire({
+      title: 'Bạn có chắc chắn?',
+      text: 'Muốn xóa collection này!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Xóa',
+      cancelButtonText: 'Hủy',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await deleteCollection(collection.id);
+        if (response?.code === 200) {
+          Swal.fire({
+            title: 'Thông báo',
+            text: response?.message,
+            icon: 'success',
+          });
+          onReload();
+        } else {
+          Swal.fire({
+            title: 'Thông báo!',
+            text: response?.message,
+            icon: 'error',
+          });
+        }
+      }
+    });
+  };
 
   return (
     <>
